+++ lib/src/generic_permission.dart
@@ -27,6 +27,7 @@ class GenericPermission with _$GenericPermission {
     GenericPermissionItem? departments,
     GenericPermissionItem? protocols,
     GenericPermissionItem? billing,
+    GenericPermissionItem? cycles,
   }) = _GenericPermission;
 
   factory GenericPermission.fromJson(Map<String, dynamic> json) => _$GenericPermissionFromJson(json);